Refactor: Move DogTrainingDTO to dto package and update references
