+++ src/main/java/org/example/dto/DogTrainingDTO.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
